requirements.txt
# Core Telegram Bot Library
python-telegram-bot==20.7

# PDF Processing Libraries
PyPDF2==3.0.1
pikepdf==8.10.1
pdf2image==1.17.0

# Image Processing
Pillow==10.2.0

# Environment Variables Management
python-dotenv==1.0.0

# System Dependencies Required (install separately):
# - poppler-utils (for pdf2image)
#   Ubuntu/Debian: sudo apt-get install poppler-utils
#   macOS: brew install poppler
#   Windows: Download from https://github.com/oschwartz10612/poppler-windows/releases/
